Refactor Streamlit plugin installation and update JacAppTest import path
